feat(hooks): consolidates logic to hook
