+++ src/components/header.tsx
@@ -1,16 +1,16 @@
 import { component$ } from '@builder.io/qwik'
-import { Link, useContent, useLocation } from '@builder.io/qwik-city'
+import { Link } from '@builder.io/qwik-city'
 
 import { Container } from './container'
-import { NavLink } from './nav-link'
-import { cx } from '~/utils/cx'
-import { isMenuItemVisible } from '~/utils/filters'
-import { ThemeToggle } from './theme-toggle'
 import { MobileNav } from './mobile-nav'
+import { NavLink } from './nav-link'
+import { ThemeToggle } from './theme-toggle'
+
+import { useMenu } from '~/hooks/use-menu'
+import { cx } from '~/utils/cx'
 
 export const Header = component$(() => {
-	const { menu } = useContent()
-	const { url } = useLocation()
+	const { options, url } = useMenu()
 
 	return (
 		<header
@@ -30,7 +30,7 @@ export const Header = component$(() => {
 				>
 					<div class='relative flex gap-4'>
 						<div class='flex flex-1'>
-							{url.pathname !== '/' && (
+							{url !== '/' && (
 								<div
 									class={cx(
 										'h-10 w-10 rounded-full bg-white/90 p-0.5 flex items-center justify-center shadow-lg shadow-slate-800/5 ring-1 ring-slate-900/5 backdrop-blur dark:bg-slate-800/90 dark:ring-white/10'
@@ -61,28 +61,24 @@ export const Header = component$(() => {
 									class='bg-white/90
                   ring-slate-900/5 flex rounded-full px-3 text-sm font-medium shadow-lg shadow-slate-800/5 ring-1 backdrop-blur dark:bg-slate-800/90 dark:ring-white/10'
 								>
-									{menu && menu.items
-										? menu.items
-												.filter(isMenuItemVisible)
-												.map(({ href, text }) => (
-													<li key={`header-link-${text}`}>
-														<NavLink
-															class={cx(
-																'relative block px-3 py-2 transition',
-																href === url.pathname
-																	? 'text-cyan-500 dark:text-cyan-400'
-																	: 'hover:text-cyan-500 dark:hover:text-cyan-400'
-															)}
-															href={href}
-														>
-															{text}
-															{url.pathname === href && (
-																<span class='absolute bg-gradient-to-r from-cyan-500 via-cyan-500/40 to-cyan-500 dark:from-cyan-400 dark:via-cyan-400/40 dark:to-cyan-400 inset-x-1 -bottom-px h-px' />
-															)}
-														</NavLink>
-													</li>
-												))
-										: null}
+									{options.value.map(({ href, text }) => (
+										<li key={`header-link-${text}`}>
+											<NavLink
+												class={cx(
+													'relative block px-3 py-2 transition',
+													href === url
+														? 'text-cyan-500 dark:text-cyan-400'
+														: 'hover:text-cyan-500 dark:hover:text-cyan-400'
+												)}
+												href={href}
+											>
+												{text}
+												{url === href && (
+													<span class='absolute bg-gradient-to-r from-cyan-500 via-cyan-500/40 to-cyan-500 dark:from-cyan-400 dark:via-cyan-400/40 dark:to-cyan-400 inset-x-1 -bottom-px h-px' />
+												)}
+											</NavLink>
+										</li>
+									))}
 								</ul>
 							</nav>
 						</div>
